# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "توثيق API منصة توبامين التعليمية",
    "version": "1.0.0",
...
    "contact": {
      "name": "فريق تطوير توبامين",
      "email": "support@topamun.com",
      "url": "https://topamun.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://topamun-backend.vercel.app/api/v1",
      "description": "خادم الإنتاج"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "خادم التطوير المحلي"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أدخل الرمز المميز الذي حصلت عليه من تسجيل الدخول"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "معرف المستخدم"
          },
          "firstName": {
            "type": "string",
            "description": "الاسم الأول"
          },
          "lastName": {
            "type": "string",
            "description": "الاسم الأخير"
          },
          "email": {
            "type": "string",
            "description": "البريد الإلكتروني"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "instructor",
              "admin"
            ],
            "description": "دور المستخدم"
          },
          "governorate": {
            "type": "string",
            "description": "المحافظة"
          },
          "gradeLevel": {
            "type": "string",
            "description": "المرحلة الدراسية (للطلاب)"
          },
          "subject": {
            "type": "string",
            "description": "المادة الدراسية (للمعلمين)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "حالة الحساب (للمعلمين)"
          },
          "isConfirmed": {
            "type": "boolean",
            "description": "هل تم تفعيل البريد الإلكتروني"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "تاريخ إنشاء الحساب"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "تاريخ آخر تحديث"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/profile": {
      "get": {
        "summary": "الحصول على الملف الشخصي",
        "description": "الحصول على بيانات الملف الشخصي للمستخدم الحالي",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم الحصول على الملف الشخصي بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم الحصول على الملف الشخصي بنجاح"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "غير مصرح لك بالوصول"
          },
          "404": {
            "description": "المستخدم غير موجود"
          }
        }
      },
      "patch": {
        "summary": "تحديث الملف الشخصي",
        "description": "تحديث بيانات الملف الشخصي للمستخدم الحالي",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "description": "الاسم الأول",
                    "example": "أحمد"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "الاسم الأخير",
                    "example": "محمد"
                  },
                  "phone": {
                    "type": "string",
                    "description": "رقم الهاتف",
                    "example": "01234567890"
                  },
                  "governorate": {
                    "type": "string",
                    "description": "المحافظة",
                    "example": "القاهرة"
                  },
                  "gradeLevel": {
                    "type": "string",
                    "description": "المرحلة الدراسية (للطلاب فقط)",
                    "example": "المرحلة الثانوية"
                  },
                  "subject": {
                    "type": "string",
                    "description": "المادة الدراسية (للمعلمين فقط)",
                    "example": "الرياضيات"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تحديث الملف الشخصي بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم تحديث الملف الشخصي بنجاح"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "خطأ في البيانات المدخلة"
          },
          "401": {
            "description": "غير مصرح لك بالوصول"
          }
        }
      }
    },
    "/api/v1/users/change-password": {
      "patch": {
        "summary": "تغيير كلمة المرور",
        "description": "تغيير كلمة المرور للمستخدم الحالي",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword",
                  "confirmNewPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "كلمة المرور الحالية",
                    "example": "CurrentPass123"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "كلمة المرور الجديدة",
                    "example": "NewPass123"
                  },
                  "confirmNewPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "تأكيد كلمة المرور الجديدة",
                    "example": "NewPass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تغيير كلمة المرور بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم تغيير كلمة المرور بنجاح"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "خطأ في البيانات المدخلة"
          },
          "401": {
            "description": "غير مصرح لك بالوصول"
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "summary": "الحصول على بيانات مستخدم محدد",
        "description": "الحصول على بيانات مستخدم محدد (للمديرين فقط)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المستخدم"
          }
        ],
        "responses": {
          "200": {
            "description": "تم الحصول على بيانات المستخدم بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم الحصول على بيانات المستخدم بنجاح"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "غير مصرح لك بالوصول"
          },
          "403": {
            "description": "ليس لديك صلاحية للوصول"
          },
          "404": {
            "description": "المستخدم غير موجود"
          }
        }
      },
      "delete": {
        "summary": "حذف مستخدم",
        "description": "حذف حساب مستخدم (للمديرين فقط)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المستخدم"
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف المستخدم بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم حذف المستخدم بنجاح"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "غير مصرح لك بالوصول"
          },
          "403": {
            "description": "ليس لديك صلاحية للوصول"
          },
          "404": {
            "description": "المستخدم غير موجود"
          }
        }
      }
    },
...
  }
}